+++ tests/test_web.py
@@ -5,13 +5,9 @@ from visual_graph_datasets.config import Config
 from visual_graph_datasets.web import NextcloudFileShare
 
 
-def test_download_nextcloud_file():
+def test_download_mock_dataset_works():
     with tempfile.TemporaryDirectory() as path:
-        file_share = NextcloudFileShare(Config())
-        file_share.url = 'https://nextcloud.electronic-heart.com/index.php/s/X49fB3LE4i6aakm'
-        file_share.download_file('metadata.md', path)
-        print(os.listdir(path))
-        with open(os.path.join(path, 'metadata.md')) as file:
-            print(file.read())
-
-        file_share.download_dataset('rb_motifs', path)
+        config = Config()
+        config.load()
+        file_share = NextcloudFileShare(config)
+        file_share.download_dataset('mock', path)