+++ tests/util.py
@@ -1,0 +1,6 @@
+import os
+import pathlib
+
+PATH = pathlib.Path(__file__).parent.absolute()
+ASSETS_PATH = os.path.join(PATH, 'assets')
+ARTIFACTS_PATH = os.path.join(PATH, 'artifacts')